Move mongo routes to routes folder
